refactor(swift): introduce AppServices container for centralized service ownership
Move six app-lifetime service instances (DaemonClient, AmbientAgent,
SurfaceManager, ToolConfirmationManager, SecretPromptManager, ZoomManager)
from scattered stored properties on AppDelegate into a single AppServices
container. AppDelegate and MainWindow use private forwarding accessors to
keep internal references working while ownership lives in the container.
Injection points (MainWindow init, VellumAssistantApp, Settings) now
access services through the container explicitly.
