+++ clients/macos/vellum-assistant/Features/MainWindow/MainWindow.swift
@@ -5,14 +5,18 @@ import SwiftUI
 
 @MainActor
 final class MainWindow {
-    private let daemonClient: DaemonClient
-    private let surfaceManager: SurfaceManager
-    private let ambientAgent: AmbientAgent
-    private let zoomManager: ZoomManager
+    private let services: AppServices
     private var window: NSWindow?
     let threadManager: ThreadManager
     let traceStore = TraceStore()
     var onMicrophoneToggle: (() -> Void)?
+
+    // Forwarding accessors — keeps existing references working while
+    // ownership lives in the `services` container.
+    private var daemonClient: DaemonClient { services.daemonClient }
+    private var surfaceManager: SurfaceManager { services.surfaceManager }
+    private var ambientAgent: AmbientAgent { services.ambientAgent }
+    private var zoomManager: ZoomManager { services.zoomManager }
 
     /// Tracks daemon reconnects so trace state can be reset on stream restart.
     private var connectionCancellable: AnyCancellable?
@@ -30,13 +34,10 @@ final class MainWindow {
         threadManager.activeViewModel
     }
 
-    init(daemonClient: DaemonClient, surfaceManager: SurfaceManager, ambientAgent: AmbientAgent, zoomManager: ZoomManager) {
-        self.daemonClient = daemonClient
-        self.surfaceManager = surfaceManager
-        self.ambientAgent = ambientAgent
-        self.zoomManager = zoomManager
-        self.threadManager = ThreadManager(daemonClient: daemonClient)
-        daemonClient.onTraceEvent = { [weak self] msg in
+    init(services: AppServices) {
+        self.services = services
+        self.threadManager = ThreadManager(daemonClient: services.daemonClient)
+        services.daemonClient.onTraceEvent = { [weak self] msg in
             Task { @MainActor in
                 self?.traceStore.ingest(msg)
             }
